Use the canonical cast names PHP 8.5 asks for (#3312)
PHP 8.5 deprecates the non-canonical casts and reports every one of
them at compile time, so the notices pile up even when the code around
them never runs. Rename them to the canonical spelling: (double) to
(float), (boolean) to (bool) and (integer) to (int). Both spellings
compile to the same opcode, so the behaviour stays the same.

The sales report writes "default;" instead of "default:", which PHP 8.5
deprecates as well. Give the label its colon.

Leave the third-party code under includes/external/ and the magnalister
files untouched, since both are maintained elsewhere.

+++ admin/includes/classes/categories.php
@@ -1600,7 +1600,7 @@ class categories {
       if (!isset($products_data['products_tax_class_id']) 
           || (int)$products_data['products_tax_class_id'] <= 0
           || !isset($products_data['products_price']) 
-          || (double)$products_data['products_price'] <= 0.00
+          || (float)$products_data['products_price'] <= 0.00
           )
       {
         $price_query = xtc_db_query("SELECT products_tax_class_id,
@@ -1612,7 +1612,7 @@ class categories {
         if (!isset($products_data['products_tax_class_id']) || (int)$products_data['products_tax_class_id'] <= 0) {
           $products_data['products_tax_class_id'] = $price['products_tax_class_id'];
         }
-        if (!isset($products_data['products_price']) || (double)$products_data['products_price'] <= 0.00) {
+        if (!isset($products_data['products_price']) || (float)$products_data['products_price'] <= 0.00) {
           $products_data['products_price'] = $price['products_price'];
         }
       }
